Move write cap checking.
