+++ kernel-ewasm/cap9-std/src/lib.rs
@@ -10,6 +10,8 @@ use validator::serialization::{Deserialize, Serialize};
 pub struct Error;
 
 pub mod proc_table;
+pub mod syscalls;
+pub use syscalls::*;
 
 // When we are compiling to WASM, unresolved references are left as (import)
 // expressions. However, under any other target symbols will have to be linked
@@ -185,119 +187,4 @@ pub fn raw_proc_write(cap_index: u8, key: &[u8; 32], value: &[u8; 32]) -> Result
     result.resize(32,0);
     // input.resize(1+1+32+32, 0);
     cap9_syscall(&input, &mut result)
-}
-
-#[derive(Clone, Debug)]
-pub struct SysCall {
-    pub cap_index: u8,
-    pub action: SysCallAction,
-}
-
-
-impl Deserialize for SysCall {
-    type Error = io::Error;
-
-    fn deserialize<R: io::Read>(reader: &mut R) -> Result<Self, Self::Error> {
-        let syscall_type = u8::deserialize(reader)?;
-        let cap_index = u8::deserialize(reader)?;
-        match syscall_type {
-            0x7 => {
-                Ok(SysCall {
-                    cap_index,
-                    action: SysCallAction::Write(WriteCall::deserialize(reader)?)
-                })
-            },
-            _ => panic!("unknown syscall"),
-        }
-    }
-}
-
-
-impl Serialize for SysCall {
-    type Error = io::Error;
-
-    fn serialize<W: io::Write>(self, writer: &mut W) -> Result<(), Self::Error> {
-        // Write syscall type
-        match self.action {
-            SysCallAction::Write(_) => writer.write(&[0x07])?
-        }
-        // Write cap index
-        writer.write(&[self.cap_index])?;
-        self.action.serialize(writer)?;
-        Ok(())
-    }
-}
-
-
-#[derive(Clone, Debug)]
-pub enum SysCallAction {
-    Write(WriteCall),
-}
-
-#[derive(Clone, Debug)]
-pub struct WriteCall {
-    pub key: U256,
-    pub value: U256,
-}
-
-impl Deserialize for WriteCall {
-    type Error = io::Error;
-
-    fn deserialize<R: io::Read>(reader: &mut R) -> Result<Self, Self::Error> {
-        let key: U256 = U256::deserialize(reader)?;
-        let value: U256 = U256::deserialize(reader)?;
-        Ok(WriteCall{key, value})
-    }
-}
-
-
-impl Serialize for SysCallAction {
-    type Error = io::Error;
-
-    fn serialize<W: io::Write>(self, writer: &mut W) -> Result<(), Self::Error> {
-        match self {
-            SysCallAction::Write(write_call) => {
-                write_call.serialize(writer)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl Serialize for WriteCall {
-    type Error = io::Error;
-
-    fn serialize<W: io::Write>(self, writer: &mut W) -> Result<(), Self::Error> {
-        // Write key
-        self.key.serialize(writer)?;
-        // Write value
-        self.value.serialize(writer)?;
-        Ok(())
-    }
-}
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pwasm_abi::types::*;
-    use validator::io;
-    use validator::serialization::{Deserialize, Serialize};
-
-    #[test]
-    fn serialize_write() {
-        let key: U256 = U256::zero();
-        let value: U256 = U256::zero();
-        let mut buffer = Vec::with_capacity(1 + 1 + 32 + 32);
-
-        let syscall = SysCall {
-            cap_index: 0,
-            action: SysCallAction::Write(WriteCall{key: key.into(), value: value.into()})
-        };
-        syscall.serialize(&mut buffer).unwrap();
-        let expected: &[u8] = &[0x7, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,0x00,0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,0x00];
-        assert_eq!(buffer, expected);
-    }
 }
